changed loadDirectory return type from int to size_t

+++ include/APluginLibrary/pluginmanager.h
@@ -42,7 +42,7 @@ namespace apl
         PluginManager& operator=(PluginManager&& other) noexcept;
 
         bool load(std::string path);
-        int loadDirectory(const std::string& path, bool recursive);
+        size_t loadDirectory(const std::string& path, bool recursive);
 
         size_t getLoadedPluginCount() const;
         std::vector<Plugin*> getLoadedPlugins();
